分辨率设置初步完成

+++ VIO/openvio_def.h
@@ -41,7 +41,7 @@ struct OPENVIO_STATUS{
 	uint8_t is_imu_send;
 
     uint8_t cam_frame_size_num;
-    uint16_t cam_frame_size;
+    uint32_t cam_frame_size;
     uint8_t cam_name;
 };
